#06 BaseRecyclerView 제거
