+++ ToyProject/app/src/main/java/com/egiwon/moviesearch/ext/RecyclerViewExt.kt
@@ -4,12 +4,12 @@ import androidx.databinding.BindingAdapter
 import androidx.paging.PagedList
 import androidx.recyclerview.widget.RecyclerView
 import com.egiwon.moviesearch.base.BaseIdentifier
-import com.egiwon.moviesearch.base.BaseRecyclerView
+import com.egiwon.moviesearch.base.BasePagedAdapter
 
 @Suppress("UNCHECKED_CAST")
 @BindingAdapter("replacePagedItems")
 fun <T : BaseIdentifier> RecyclerView.replacePagedItems(items: PagedList<T>?) {
-    (adapter as? BaseRecyclerView.BasePagedAdapter<T, *>)?.run {
+    (adapter as? BasePagedAdapter<T, *>)?.run {
         replaceAllPagedItems(items)
     }
 }
